Second Commit

+++ application/views/v_stok.php
@@ -1,3 +1,6 @@
+<?php
+  $jabatan_user = $this->session->userdata('jabatan_user');;
+?>
 				<main>
                     <div class="container-fluid">
                         <h1 class="mt-4"></h1>
@@ -20,7 +23,9 @@
                                                 <th>Stok Akhir</th>
                                                 <th>Penjualan Bulanan</th>
                                                 <th>Persediaan Bulanan</th>
+												<?php if($jabatan_user=="admin"){ ?>
                                                 <th>Action</th>
+												<?php } ?>
                                             </tr>
                                         </thead>
                                         <tbody>
@@ -36,6 +41,14 @@
 <script src="https://cdn.datatables.net/1.10.20/js/dataTables.bootstrap4.min.js" crossorigin="anonymous"></script>
 <script>
 $(document).ready(function() {
+	
+var jabatan = "<?php echo $jabatan_user;?>";
+if(jabatan=="admin"){
+	var angka = 5;
+}else{
+	var angka = 4;
+}
+
 $('#datastok').DataTable({
 			"ajax": "<?php echo base_url('stok/get_data_master_stok/');?>" ,
 			"columns": [
@@ -45,20 +58,24 @@ $('#datastok').DataTable({
 				{ "data": "penjualan_stok" },
 				{ "data": "persediaan_stok" }
 			],
+						
 			columnDefs: [
 				   {
-				   targets: [5],
+				   targets: [angka],
 				    data: 'id_stok',
 				   render: function ( data, type, row, meta ) { 
-
-					var edit = "<a href='<?php echo base_url();?>stok/stok_ubah/"+data+"' title='Ubah'> <button type='button' class='btn btn-sm btn-warning'><i class='fa fa-pencil'></i> </button></a>";
-					
-					var hapus = "<a onclick=\"return confirm('Yakin untuk menghapus stok ini ?')\" href='<?php echo base_url();?>stok/stok_aksi_hapus/"+data+"' title='Hapus'> <button type='button' class='btn btn-sm  btn-danger'><i class='fa fa-trash'></i> </button></a>";
-					
-					
-					return edit +  hapus;
+					if(jabatan=="admin"){
+						var edit = "<a href='<?php echo base_url();?>stok/stok_ubah/"+data+"' title='Ubah'> <button type='button' class='btn btn-sm btn-warning'><i class='fa fa-pencil'></i> </button></a>";
+						
+						var hapus = "<a onclick=\"return confirm('Yakin untuk menghapus stok ini ?')\" href='<?php echo base_url();?>stok/stok_aksi_hapus/"+data+"' title='Hapus'> <button type='button' class='btn btn-sm  btn-danger'><i class='fa fa-trash'></i> </button></a>";
+						
+						return edit +  hapus;
+					}else{
+						return row.persediaan_stok;
+					}
 				   }
 				},],
+				
 			});
 	});
 </script>
